requirements.txt
Django==1.10.3
djangorestframework==3.5.3
mysql-connector-python==2.0.4
mysqlclient==1.3.7

